+++ src/core/core.go
@@ -28,15 +28,16 @@ type Core struct {
 	// guarantee that it will be covered by the mutex
 	phony.Inbox
 	*iwe.PacketConn
-	ctx          context.Context
-	cancel       context.CancelFunc
-	secret       ed25519.PrivateKey
-	public       ed25519.PublicKey
-	links        links
-	proto        protoHandler
-	log          Logger
-	addPeerTimer *time.Timer
-	config       struct {
+	ctx           context.Context
+	cancel        context.CancelFunc
+	secret        ed25519.PrivateKey
+	public        ed25519.PublicKey
+	networkSecret []byte
+	links         links
+	proto         protoHandler
+	log           Logger
+	addPeerTimer  *time.Timer
+	config        struct {
 		tls *tls.Config // immutable after startup
 		//_peers             map[Peer]*linkInfo         // configurable after startup
 		_listeners         map[ListenAddress]struct{} // configurable after startup